+++ CHANGELOG.rst
@@ -4,6 +4,13 @@ Changelog
 :code:`mutatest` is alpha software, and backwards compatibility between releases is
 not guaranteed while under development.
 
+0.9.2
+-----
+
+    - Added :code:`--exception` and :code:`-x` as a suvivor tolerance to raise an exception
+      after the trial completes if the count of surviving mutants is greater than or equal to the
+      specified value.
+
 0.9.1
 -----